Check confirmation height processing queue in RPCs block_info and blocks_info when returning confirmed status
